Factored out the Electron's speaker and adjusted instance variable naming.
